Handle more cases in findElement
When an error of type 'no such element' is detected WebDriverIO has a
special handling for it and returns the error object instead of the
element reference.

With this fix we should handle both cases where this detection has
worked and cases where it hasn't but the W3C error is still 'no such
element'.